Able to pass data by calling getController
This took me a while to figure out but basically the you need to have to call 'loader.load()' first before being able to access anything in the Controller obtained by 'getController()'. Otherwise, you will end up with a null pointer exception. This is because the 'initialize()' method in a Controller runs after the 'loader.load()' is called. Before that any field in the Controller will be null as it would not have been initialised.